v0.9.6

+++ GezelimForm/LatLong.cs
@@ -8,7 +8,7 @@ namespace GezelimForm
 {
     class LatLong
     {
-        public List<float> locationsX { get; set; }
-        public List<float> locationsY { get; set; }
+        public List<double> locationsX { get; set; }
+        public List<double> locationsY { get; set; }
     }
 }
